requirements.txt
asgiref>=3.8.1
dj-database-url>=2.3.0
Django>=5.1.7
django-cors-headers>=4.7.0
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.5.0
drf-spectacular>=0.28.0
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
whitenoise>=6.9.0
PyJWT>=2.9.0
PyYAML>=6.0.2
requests>=2.32.0